sublime

+++ embarcados/trab2/wrapper/apps/codigo.java
@@ -17,12 +17,14 @@ public class codigo {
         int maxLen = 0;
         int fl = first.length();
         int sl = second.length();
-        int[][] table = new int[fl+1][sl+1];
+        short[][] table = new short[fl+1][sl+1];
     
         for (int i = 1; i <= fl; i++) {
             for (int j = 1; j <= sl; j++) {
                 if (first.charAt(i-1) == second.charAt(j-1)) {
-                        table[i][j] = table[i - 1][j - 1] + 1;
+                        table[i][j] = (short) (table[i - 1][j - 1] + 1);
+                        // if (table[i][j] > 100)
+                        //     System.out.println(table[i][j]);
                     if (table[i][j] > maxLen)
                         maxLen = table[i][j];
                 }
